Intermediate commit (adding protocol routing model)
