--- Spravo/Storyboards/Base.lproj/Main.storyboard
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="BYZ-38-t0r">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--LoginVC-->
-        <scene sceneID="tne-QT-ifu">
-            <objects>
-                <viewController id="BYZ-38-t0r" customClass="LoginVC" customModule="Spravo" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="PFo-DG-4mp">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                        <subviews>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="gQV-pW-1av" customClass="FBSDKLoginButton">
-                                <rect key="frame" x="164.5" y="318.5" width="46" height="30"/>
-                                <state key="normal" title="Button"/>
-                            </button>
-                            <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="SpravoBrand" translatesAutoresizingMaskIntoConstraints="NO" id="lg6-Dp-Zl0">
-                                <rect key="frame" x="97.5" y="222" width="180" height="80"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="80" id="2TE-92-Syh"/>
-                                    <constraint firstAttribute="width" constant="180" id="fVo-xh-kfO"/>
-                                </constraints>
-                            </imageView>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="{Label}" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Fky-zE-Wrh">
-                                <rect key="frame" x="155.5" y="181.5" width="64.5" height="25.5"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="21"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="lg6-Dp-Zl0" firstAttribute="centerX" secondItem="PFo-DG-4mp" secondAttribute="centerX" id="4hg-fs-9WV"/>
-                            <constraint firstItem="gQV-pW-1av" firstAttribute="centerX" secondItem="PFo-DG-4mp" secondAttribute="centerX" id="BLC-XR-UoL"/>
-                            <constraint firstItem="gQV-pW-1av" firstAttribute="top" secondItem="lg6-Dp-Zl0" secondAttribute="bottom" constant="16.5" id="Bop-A3-1RQ"/>
-                            <constraint firstItem="Fky-zE-Wrh" firstAttribute="centerX" secondItem="PFo-DG-4mp" secondAttribute="centerX" id="nyn-cY-u8Y"/>
-                            <constraint firstItem="lg6-Dp-Zl0" firstAttribute="top" secondItem="Fky-zE-Wrh" secondAttribute="bottom" constant="15" id="sUW-W6-850"/>
-                            <constraint firstItem="gQV-pW-1av" firstAttribute="centerY" secondItem="PFo-DG-4mp" secondAttribute="centerY" id="zpm-ia-b7I"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="lGW-rB-hzy"/>
-                    </view>
-                    <connections>
-                        <outlet property="facebookLoginButton" destination="gQV-pW-1av" id="lZ9-WC-Ldz"/>
-                        <outlet property="messageLabel" destination="Fky-zE-Wrh" id="2Nw-IR-jKP"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="56.799999999999997" y="26.53673163418291"/>
-        </scene>
-    </scenes>
-    <resources>
-        <image name="SpravoBrand" width="180" height="80"/>
-    </resources>
-</document>
